pyproject.toml
[project]
name = "achevrot-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jinja2>=3.1.6",
    "markdown>=3.10",
    "pygments>=2.19.2",
    "pyyaml>=6.0.3",
]
